# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Golang Quickstart using Gin Gonic",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/airport/direct-connections": {
            "get": {
                "description": "Get Direct Connections from specified Airport",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airport collection"
                ],
                "summary": "Get Direct Connections from Airport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Source airport\u003cbr\u003eExample: SFO, LHR, CDG",
                        "name": "airport",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of direct connections to return (page size)\u003cbr\u003eDefault value : 10",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of direct connections to skip (for pagination)\u003cbr\u003eDefault value : 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Destination"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/airport/list": {
            "get": {
                "description": "Get list of Airports. Optionally, you can filter the list by Country",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airport collection"
                ],
                "summary": "List Airport Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country\u003cbr\u003eExample: United Kingdom, France, United States",
                        "name": "country",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of airports to return (page size)\u003cbr\u003eDefault value : 10",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of airports to skip (for pagination)\u003cbr\u003eDefault value : 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Airport"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/airport/{id}": {
            "get": {
                "description": "Get Airport with specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airport collection"
                ],
                "summary": "Get Airport Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airport ID like airport_1273",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Airport"
                        }
                    },
                    "404": {
                        "description": "Airport Document ID Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "Update Airport with specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airport collection"
                ],
                "summary": "Update Airport Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airport ID like airport_1273",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updates document",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Airport"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Airport"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Create Airport with specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airport collection"
                ],
                "summary": "Insert Airport Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airport ID like airport_1273",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data to create a document",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Airport"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Airport"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "409": {
                        "description": "Airport Document already exists"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "Delete Airport with specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airport collection"
                ],
                "summary": "Deletes Airport Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airport ID like airport_1273",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Airport deleted"
                    },
                    "404": {
                        "description": "Airport Document ID Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Airport": {
            "type": "object",
            "required": [
                "airportname",
                "city",
                "country",
                "faa"
            ],
            "properties": {
                "airportname": {
                    "type": "string",
                    "example": "SampleAirport"
                },
                "city": {
                    "type": "string",
                    "example": "SampleCity"
                },
                "country": {
                    "type": "string",
                    "example": "United Kingdom"
                },
                "faa": {
                    "type": "string",
                    "example": "SAA"
                },
                "geo": {
                    "$ref": "#/definitions/models.Geo"
                },
                "icao": {
                    "type": "string",
                    "example": "SAAA"
                },
                "tz": {
                    "type": "string",
                    "example": "Europe/Paris"
                }
            }
        },
        "models.Destination": {
            "type": "object",
            "properties": {
                "destinationairport": {
                    "type": "string",
                    "example": "JFK"
                }
            }
        },
        "models.Geo": {
            "type": "object",
            "properties": {
                "alt": {
                    "type": "number",
                    "example": 48.864716
                },
                "lat": {
                    "type": "number",
                    "example": 2.349014
                },
                "lon": {
                    "type": "number",
                    "example": 92
                }
            }
        },
...
    }
}